EXTREME MONOREPO REFACTOR!!!
Moved the mobile-controller from the KNIRVENGINE into the KNIRVCONTROLLER and renamed it to manager.
Moved KNIRVSHELL from the project root and placed it into KNIRVCONTROLLER as shell.
Cloned agent-core from KNIRVENGINE into KNIRVCONTROLLER, as trainer, for the frontend capability.
Cloned agent-core from KNIRVENGINE into KNIRVCORTEX, in the subdirectory agent-core in hopes of isolating the backend WASM compilation pipeline, model training, and deployment sequence of the WASM agent core build file.